6.3 Создаст компонент Cities, подключит его к хранилищу вместо OffersList
